+++ web-interface/styles/OpenedActivityCard.module.css
@@ -1,4 +1,4 @@
-div.Card{
+div.Card {
     box-sizing: border-box;
     display: flex;
     width: 22em;
@@ -13,12 +13,12 @@ div.Card{
 }
 
 
-div.close{
+div.close {
     display: flex;
     justify-content: flex-end;
 }
 
-div.title{
+div.title {
     width: 100%;
     position: relative;
     padding-left: 2em;
@@ -29,33 +29,33 @@ div.title{
 
 }
 
-div.ownersWrap{
+div.ownersWrap {
     display: flex;
     justify-content: space-between;
     padding-top: 1.5em;
     font-family: 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
 }
 
-div.owner{
+div.owner {
     padding-left: 2em;
     justify-content: center;
     text-align: center;
 }
 
-img.ownerPhoto{
+img.ownerPhoto {
     border-radius: 50%;
     width: 4em;
     height: 4em;
 }
 
-div.coOwners{
+div.coOwners {
     padding-right: 4.5em;
     position: relative;
     top: 0;
     left: 0;
 }
 
-img.image1{
+img.image1 {
     border-radius: 50%;
     width: 4em;
     height: 4em;
@@ -67,7 +67,7 @@ img.image1{
 }
 
 
-img.image2{
+img.image2 {
     border-radius: 50%;
     width: 4em;
     height: 4em;
@@ -78,7 +78,7 @@ img.image2{
     z-index: 1;
 }
 
-img.image3{
+img.image3 {
     border-radius: 50%;
     width: 4em;
     height: 4em;
@@ -90,7 +90,7 @@ img.image3{
 }
 
 
-div.image1{
+div.image1 {
     border-radius: 50%;
     width: 4em;
     height: 4em;
@@ -101,7 +101,7 @@ div.image1{
     z-index: 2;
 }
 
-div.image2{
+div.image2 {
     border-radius: 50%;
     width: 4em;
     height: 4em;
@@ -112,7 +112,7 @@ div.image2{
     z-index: 1;
 }
 
-div.image3{
+div.image3 {
     border-radius: 50%;
     width: 4em;
     height: 4em;
@@ -123,21 +123,21 @@ div.image3{
     z-index: 0;
 }
 
-div.ownerPhoto{
+div.ownerPhoto {
     border-radius: 50%;
     width: 4em;
     height: 4em;
 }
 
 
-div.letter{
+div.letter {
     color: white;
     text-align: center;
     padding-top: 0.2em;
     font-size: 2em;
 }
 
-div.description{
+div.description {
     padding-left: 2em;
     padding-right: 2em;
     padding-top: 1em;
@@ -146,7 +146,7 @@ div.description{
     color: #505050;
 }
 
-div.addComment{
+div.addComment {
     display: flex;
     justify-content: center;
     padding-top: 2em;
@@ -154,7 +154,7 @@ div.addComment{
 }
 
 
-input.inputComment{
+input.inputComment {
     width: 14em;
     border: none;
     border-radius: 0px;
@@ -169,49 +169,49 @@ input.inputComment{
 
 }
 
-button.sendButton{
+button.sendButton {
     background-color: #eaeaea;
     border: 1px solid #eaeaea;
 }
 
-img.send{
-    background-color:#eaeaea;
+img.send {
+    background-color: #eaeaea;
     width: 2em;
 }
 
-.rotated{
-    margin-left: 0.4em ;
+.rotated {
+    margin-left: 0.4em;
     transform: rotate(180deg);
 }
 
-.nonRotated{
+.nonRotated {
     padding-left: 0.4em;
     transform: rotate(0);
 }
 
 
-div.commentsWrap{
+div.commentsWrap {
     display: flex;
     color: #505050;
 }
 
-div.commentsText{
+div.commentsText {
     font-family: 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
     text-align: left;
 
     padding-left: 3em;
     padding-top: 2em;
-    
+
 }
 
-div.ownerExistsButNoPhoto{
+div.ownerExistsButNoPhoto {
     background-color: var(--dynamic-owner-);
     border-radius: 50%;
     width: 4em;
     height: 4em;
 }
 
-button.closeButton{
+button.closeButton {
     background-color: white;
     border: none;
     margin-top: 1em;
@@ -228,38 +228,47 @@ div.modalBackground {
     justify-content: center;
     align-items: center;
 
-    background-color:rgba(0, 0, 0, 0.5);
+    background-color: rgba(0, 0, 0, 0.5);
     backdrop-filter: blur(0.5em);
 }
 
-div.commentsContainer{
+div.commentsContainer {
     margin-top: 1em;
     height: 10.5em;
     align-self: center;
     display: var(--comments-display-);
-    overflow:scroll;
+    overflow: scroll;
     overflow-x: hidden;
 
 }
 
-button.cameraIcon{
+div.cameraIcon {
     border: none;
     background-color: white;
     margin: 0.2em;
 }
 
-div.scroll{
+div.cameraIcon label {
+    margin-top: 0.5em;
+    margin-right: 0.5em;
+}
+
+div.cameraIcon>input {
+    display: none;
+}
+
+div.scroll {
     height: var(--dynamic-opened-card-height-);
     /* height: 40em; */
-    overflow-y:scroll;
+    overflow-y: scroll;
     overflow-x: hidden;
 }
 
-div.nonScroll{
+div.nonScroll {
     height: auto;
 }
 
-button.arrowButton{
+button.arrowButton {
     margin-top: 1.8em;
     border: none;
     background-color: white;
@@ -267,29 +276,34 @@ button.arrowButton{
 }
 
 @media screen and (max-width: 380px) {
-    div.Card{
+    div.Card {
         width: 18em;
     }
-    input.inputComment{
+
+    input.inputComment {
         width: 10.5em;
     }
-    div.coOwners{
+
+    div.coOwners {
         padding-right: 4em;
     }
 }
 
 
 @media screen and (max-width: 300px) {
-    div.Card{
+    div.Card {
         width: 16em;
     }
-    input.inputComment{
+
+    input.inputComment {
         width: 8.5em;
     }
-    div.coOwners{
+
+    div.coOwners {
         padding-right: 3.5em;
     }
-    div.owner{
+
+    div.owner {
         padding-left: 1em;
     }
 }
